+++ db/boinc_db_types.h
@@ -220,7 +220,7 @@ struct USER {
 struct USER_DELETED {
     DB_ID_TYPE userid;
     char public_cross_project_id[256];
-    int create_time;
+    double create_time;
     void clear();
 };
 
@@ -389,7 +389,7 @@ struct HOST {
 struct HOST_DELETED {
     DB_ID_TYPE hostid;
     char public_cross_project_id[256];
-    int create_time;
+    double create_time;
     void clear();
 };
